--- app/src/main/java/com/techcourse/ManualHandlerMapping.java
@@ -1,52 +1,0 @@
-package com.techcourse;
-
-import com.techcourse.controller.manual_controller.LoginController;
-import com.techcourse.controller.manual_controller.LoginViewController;
-import com.techcourse.controller.manual_controller.LogoutController;
-import com.techcourse.controller.manual_controller.RegisterController;
-import com.techcourse.controller.manual_controller.RegisterViewController;
-import jakarta.servlet.http.HttpServletRequest;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import webmvc.org.springframework.web.servlet.mvc.HandlerMapping;
-import webmvc.org.springframework.web.servlet.mvc.asis.Controller;
-import webmvc.org.springframework.web.servlet.mvc.asis.ForwardController;
-
-import java.util.HashMap;
-import java.util.Map;
-
-public class ManualHandlerMapping implements HandlerMapping {
-
-    private static final Logger log = LoggerFactory.getLogger(ManualHandlerMapping.class);
-
-    private static final Map<String, Controller> controllers = new HashMap<>();
-
-    @Override
-    public void initialize() {
-        controllers.put("/", new ForwardController("/index.jsp"));
-        controllers.put("/login", new LoginController());
-        controllers.put("/login/view", new LoginViewController());
-        controllers.put("/logout", new LogoutController());
-        controllers.put("/register/view", new RegisterViewController());
-        controllers.put("/register", new RegisterController());
-
-        log.info("Initialized Handler Mapping!");
-        controllers.keySet()
-                .forEach(path -> log.info("Path : {}, Controller : {}", path, controllers.get(path).getClass()));
-    }
-
-    @Override
-    public boolean containsHandler(final HttpServletRequest request) {
-        return controllers.containsKey(request.getRequestURI());
-    }
-
-    @Override
-    public Object getHandler(final HttpServletRequest request) {
-        return getHandler(request.getRequestURI());
-    }
-
-    public Controller getHandler(final String requestURI) {
-        log.debug("Request Mapping Uri : {}", requestURI);
-        return controllers.get(requestURI);
-    }
-}
