--- app/src/main/java/com/techcourse/controller/manual_controller/RegisterController.java
@@ -1,21 +1,0 @@
-package com.techcourse.controller.manual_controller;
-
-import com.techcourse.domain.User;
-import com.techcourse.repository.InMemoryUserRepository;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import webmvc.org.springframework.web.servlet.mvc.asis.Controller;
-
-public class RegisterController implements Controller {
-
-    @Override
-    public String execute(final HttpServletRequest req, final HttpServletResponse res) {
-        final var user = new User(2,
-                req.getParameter("account"),
-                req.getParameter("password"),
-                req.getParameter("email"));
-        InMemoryUserRepository.save(user);
-
-        return "redirect:/index.jsp";
-    }
-}
