--- app/src/main/java/com/techcourse/controller/manual_controller/RegisterViewController.java
@@ -1,13 +1,0 @@
-package com.techcourse.controller.manual_controller;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import webmvc.org.springframework.web.servlet.mvc.asis.Controller;
-
-public class RegisterViewController implements Controller {
-
-    @Override
-    public String execute(final HttpServletRequest req, final HttpServletResponse res) {
-        return "/register.jsp";
-    }
-}
