--- mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/asis/ControllerHandlerAdapter.java
@@ -1,26 +1,0 @@
-package webmvc.org.springframework.web.servlet.mvc.asis;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import webmvc.org.springframework.web.servlet.ModelAndView;
-import webmvc.org.springframework.web.servlet.mvc.HandlerAdapter;
-import webmvc.org.springframework.web.servlet.view.JspView;
-
-public class ControllerHandlerAdapter implements HandlerAdapter {
-
-    @Override
-    public boolean supports(final Object handler) {
-        return handler instanceof Controller;
-    }
-
-    @Override
-    public ModelAndView handle(
-            final HttpServletRequest request,
-            final HttpServletResponse response,
-            final Object handler
-    ) throws Exception {
-        final String viewName = ((Controller) handler).execute(request, response);
-
-        return new ModelAndView(new JspView(viewName));
-    }
-}
